requirements.txt
Django>=4.2
channels>=4.0
channels_redis>=4.0
daphne>=4.0
asgiref>=3.0
redis>=4.0
djangorestframework
django-cors-headers
